requirements.txt
tinker-api
playwright
Pillow
scikit-image
numpy
torch
tqdm
datasets
